+++ indra/media_plugins/CMakeLists.txt
@@ -3,7 +3,7 @@
 add_subdirectory(base)
 
 if (LINUX)
-    #add_subdirectory(gstreamer010)
+    add_subdirectory(gstreamer10)
     add_subdirectory(example)
 endif (LINUX)
