+++ indra/media_plugins/base/media_plugin_base.h
@@ -32,6 +32,40 @@
 #include "llpluginmessage.h"
 #include "llpluginmessageclasses.h"
 
+#if LL_LINUX
+
+struct SymbolToGrab
+{
+    bool mRequired;
+    char const *mName;
+    void **mPPFunc;
+};
+
+class SymbolGrabber
+{
+public:
+    size_t registerSymbol( SymbolToGrab aSymbol );
+    bool grabSymbols(std::vector< std::string > const &aDSONames);
+    void ungrabSymbols();
+
+private:
+    std::vector< SymbolToGrab > gSymbolsToGrab;
+
+    bool sSymsGrabbed = false;
+    std::vector<void *> sLoadedLibraries;
+};
+
+extern SymbolGrabber gSymbolGrabber;
+
+// extern SymbolGrabber gSymbolGrabber;
+
+#define LL_GRAB_SYM(SYMBOL_GRABBER, REQUIRED, SYMBOL_NAME, RETURN, ...) \
+    RETURN (*ll##SYMBOL_NAME)(__VA_ARGS__) = nullptr; \
+    size_t gRegistered##SYMBOL_NAME = SYMBOL_GRABBER.registerSymbol( \
+        { REQUIRED, #SYMBOL_NAME , (void**)&ll##SYMBOL_NAME} \
+    );
+
+#endif
 
 class MediaPluginBase
 {
@@ -126,4 +160,7 @@ int init_media_plugin(
     LLPluginInstance::sendMessageFunction *plugin_send_func,
     void **plugin_user_data);
 
-
+#if LL_LINUX
+pid_t getParentPid(pid_t aPid);
+bool isPluginPid(pid_t aPid);
+#endif
