--- indra/media_plugins/gstreamer010/llmediaimplgstreamer_syms_rawv.inc
@@ -1,5 +1,0 @@
-
-// required symbols to grab
-LL_GST_SYM(true, gst_video_sink_get_type, GType, void);
-
-// optional symbols to grab
